api/requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
python-dotenv==0.19.1
waitress==2.0.0
Werkzeug==2.0.2
